Students controlller with add student functionality of teacher

+++ database/migrations/2021_02_17_052220_create_teachers_students_table.php
@@ -16,7 +16,7 @@ class CreateTeachersStudentsTable extends Migration
         Schema::create('teachers_students', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('student_id');
-            $table->unsignedInteger('teacher_id');
+            $table->unsignedBigInteger('teacher_id');
             $table->foreign('student_id')->references('id')->on('students');
             $table->foreign('teacher_id')->references('id')->on('teachers');
             $table->timestamps();
